+++ doleus/utils/utils.py
@@ -1,3 +1,6 @@
+# SPDX-FileCopyrightText: 2025 Doleus contributors
+# SPDX-License-Identifier: Apache-2.0
+
 import datetime
 from typing import Any, Union