+++ tests/utils/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2025 Doleus contributors
+# SPDX-License-Identifier: Apache-2.0
+